skills-based, uniform diet and meal cooked
Friendship links are black, and thus not visible
In current model version, each person is assigned cooking skills based on a random distribution that has a maximum value for each diet. It is possible in this version that a person’s dietary preference is for example meat, but they can only cook fish.
Used meat, fish, vegetarian, vegan throughout the whole model to account for diet and meal cooked.
Created skills-based procedure
Fixed bug of empty household