added shops gradually more sustainable
Added a procedure that food outlets gradually increase their vegetarian and vegan stock. For this, in setup of food outlets I added vegetarian and vegan to initial stock table, stock table, and sales table as well, with initial values of 0.
I added the option of asking random persons to change their diet in the procedure for diet-influencers.