updated change plant and animal protein procedures
Changes names from sustainable and less, to change in plant and animal protein
Restructured the procedure for updating the stocks based on assortment change (pos or neg) and if the food outlet already offers and/or no longer offers a product.
Added code to have the food outlets update their label.